+++ changelog.md
@@ -1,11 +1,13 @@
 # Changelog
 
-## Version 8.0.0 (2020-08-10)
+## Version 8.0.0 (2020-09-08)
 
 ### Features
 
 - added new `AuthInterceptorOptions` as options for `AuthInterceptor`
     - new option `treatUnauthorizedAsForbidden`, which allows treating *401* as *403* http code
+- added new property `userIdentity`, storing last value of `UserIdentity`
+- added new method `isAuthorizedSync`, used for synchronous checking whether user has *permission*
 
 ### BREAKING CHANGES
 
@@ -13,6 +15,7 @@
 - minimal supported version of `@jscrpt/common` is `1.2.0`
 - minimal supported version of `@anglr/common` is `8.0.0`
 - removed `AuthInterceptorConfig`, now using `AuthenticationService` instead
+- `AuthorizeDirective` now initialy working synchronously
 - `AuthInterceptor` has new constructor parameters
     - default behavior of `AuthInterceptor` is not to treat *401* as *403* http code as was before